Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -19,7 +19,7 @@ public:
         
     }
     bool Ispossible(int mid,vector<int>&piles,int h){
-        int totalhours=0;
+        long long totalhours=0;
         for(auto &pile:piles){
             totalhours+=(pile+mid-1)/mid;
         }
